point based utility function
Added POINTS look-up table to Piece class.

Added tally_points() static method to Khet class.

Added `over` flag to game state. Using the points method this is how the terminal state is
determined from an arbitrary state.

The fire_laser() function now returns whether or not a pharoah has been hit. In result()
the points are tallied for a given state and stored as the utility within the game state.