laser traverses board
I still need to account for some edge cases. The current laser
implementation is not at all the fastest I can think of, but it is
serving as a crude model that works for the time being. Success with the
laser will help tune the movegen performance test even though now I will
have to tune both the move generation and laser individually.

The laser needs to return a discrete result that can be fed to the game
board such that the board quickly resolves the laser hit.

Some re-architecture for the game engine may be in order. In both move
generation and here (the laser) I am seeing the usefulness of a piece
tracking array for quick look-up of what boards are involved for what
squares. Early on I used piece codes, including including earlier in the
older Python implementation. They're pretty useful since Khet pieces
have three stateful properties.